Provide polyfill shim for torch.no_grad() (#644)
