+++ pyro/infer/trace_elbo.py
@@ -59,8 +59,6 @@ class Trace_ELBO(object):
         """
         runs the guide and runs the model against the guide with
         the result packaged as a trace generator
-
-        XXX support for automatically settings args/kwargs to volatile?
         """
 
         for i in range(self.num_particles):
